feat(discriminator): drop a leak-vector feature, fix a float32 overflow, add the case study
Removed `valid_channel_fraction`. It looks like a harmless quality flag and is not one: how
many receivers a window has tracks network density, which tracks region and epoch, so a
model could use it to infer where and when an event happened.

Fixed a float32 overflow: the finiteness check ran on the float64 array, so a value above
float32 max became inf on the cast and entered the store as a silently poisoned trace. One
window in 1925 was affected. The fix changed the trained model, so the test set was scored
a second time; seal_version is bumped to 2 and the first, superseded result is kept under
reports/m1/superseded/ so the reader can see the numbers moved slightly DOWN, not up.

Wired the ComCat landslide fixture into the positive join, which the CLI had never passed
even though `assemble` supports it. The built dataset predates this fix and does not
contain those 57 events; that is recorded as a gap rather than papered over.

Adds `serac models case-study`, which scores an event excluded for thin receiver coverage
with the already-sealed model and labels it a single-window case study, never a metric.
The alternative — lowering the receiver threshold after finding it excludes the headline
event — would be exactly the post-hoc tuning this component exists to avoid.

+++ src/serac/models/discriminator/windows.py
@@ -292,9 +292,12 @@ def process_station_window(
         data = np.asarray(trace.data, dtype=np.float64)
         if data.size < N_SAMPLES:
             data = np.pad(data, (0, N_SAMPLES - data.size))
-        sliced = data[:N_SAMPLES]
+        # Check finiteness *after* the float32 cast: response removal near a sensor's
+        # low-frequency corner can produce float64 values above float32 max, which become inf
+        # on the cast and would enter the store as a silently poisoned trace.
+        sliced = data[:N_SAMPLES].astype(np.float32)
         if not np.all(np.isfinite(sliced)) or float(np.abs(sliced).max()) == 0.0:
             continue
-        out[index] = sliced.astype(np.float32)
+        out[index] = sliced
         valid[index] = True
     return out, valid
